backend/requirements.txt
fastapi
uvicorn[standard]
pytesseract
pdf2image
Pillow
numpy
python-dotenv
pydantic
python-multipart
rapidfuzz
google-generativeai
supabase
crewai
